Move YieldProcessorNormalized into separate files
